+++ tests/tests_house.py
@@ -1,35 +1,34 @@
-import datetime
-import json
-
 from tests.tests_tibiapy import TestCommons
 from tibiapy import InvalidContent
-from tibiapy.builders import HouseBuilder
 from tibiapy.enums import HouseOrder, HouseStatus, HouseType
 from tibiapy.models import House, HouseEntry, HousesSection
 from tibiapy.parsers import HouseParser, HousesSectionParser
 from tibiapy.urls import get_house_url
 
 FILE_HOUSE_RENTED = "house/houseRented.txt"
-FILE_HOUSE_STATUS_TRANSFER = "house/houseStatusTransferred.txt"
-FILE_HOUSE_STATUS_NO_BIDS = "house/houseStatusNoBids.txt"
-FILE_HOUSE_STATUS_WITH_BIDS = "house/houseAuctionedWithBids.txt"
-FILE_HOUSE_STATUS_RENTED = "house/houseStatusRented.txt"
+FILE_HOUSE_RENTED_ACCEPTED_TRANSFER = "house/houseRentedAcceptedTransfer.txt"
+FILE_HOUSE_AUCTION_WITH_BIDS = "house/houseAuctionedWithBids.txt"
+FILE_HOUSE_AUCTION_WITHOUT_BIDS = "house/houseAuctionedWithoutBids.txt"
 FILE_HOUSE_NOT_FOUND = "house/houseNotFound.txt"
+
 FILE_HOUSE_LIST = "housesSection/housesSection.txt"
 FILE_HOUSE_LIST_NOT_FOUND = "housesSection/housesSectionNotFound.txt"
 FILE_HOUSE_LIST_EMPTY = "housesSection/housesSectionEmpty.txt"
+FILE_HOUSE_SECTION_AUCTIONED_HOUSES = "housesSection/housesSectionWithAuctionedHouses.txt"
 
 
 class TestsHouse(TestCommons):
+
+    # region House Tests
     def test_house_parser_from_content(self):
         """Testing parsing a house"""
         content = self.load_resource(FILE_HOUSE_RENTED)
         house = HouseParser.from_content(content)
 
         self.assertIsInstance(house, House)
-        self.assertEqual("Sorcerer's Avenue Labs 2c", house.name)
-        self.assertEqual(2, house.beds)
-        self.assertTrue(50_000, house.rent)
+        self.assertIsNotNone(house.name)
+        self.assertIsInstance(house.beds, int)
+        self.assertIsInstance(house.rent, int)
         self.assertEqual(HouseStatus.RENTED, house.status)
         self.assertEqual(house.url, get_house_url(house.world, house.id))
         self.assertIsNotNone(house.owner)
@@ -37,50 +36,46 @@ class TestsHouse(TestCommons):
         self.assertIsNone(house.highest_bidder)
         self.assertIsNone(house.highest_bidder_url)
 
-        house_json_raw = house.model_dump_json()
-        house_json = json.loads(house_json_raw)
-        self.assertEqual(house.image_url, house_json["image_url"])
-
-    def test_house_parser_from_content_transferred(self):
-        """Testing parsing a house being transferred"""
-        house = HouseBuilder().name("Name")
-        content = self.load_resource(FILE_HOUSE_STATUS_TRANSFER)
-        HouseParser._parse_status(house, content)
-        self.assertEqual(house._status, HouseStatus.RENTED)
-        self.assertEqual(house._owner, "Xenaris mag")
-        self.assertEqual(house._transfer_recipient, "Ivarr Bezkosci")
-        self.assertTrue(house._transfer_accepted)
-        self.assertEqual(house._transfer_price, 850000)
-
-    def test_house_parse_status_rented(self):
-        """Testing parsing a rented status"""
-        house = HouseBuilder().name("Name")
-        content = self.load_resource(FILE_HOUSE_STATUS_RENTED)
-        HouseParser._parse_status(house, content)
-        self.assertEqual(house._status, HouseStatus.RENTED)
-        self.assertEqual(house._owner, "Thorcen")
-        self.assertIsInstance(house._paid_until, datetime.datetime)
-
-    def test_house_parser_from_content_house_with_bids(self):
-        """Testing parsing a house status with bids"""
-        content = self.load_resource(FILE_HOUSE_STATUS_WITH_BIDS)
+    def test_house_parser_from_content_rented_accepted_transfer(self):
+        content = self.load_resource(FILE_HOUSE_RENTED_ACCEPTED_TRANSFER)
 
         house = HouseParser.from_content(content)
 
+        self.assertIsNotNone(house)
+        self.assertEqual(HouseStatus.RENTED, house.status)
+        self.assertIsNotNone(house.owner)
+        self.assertIsNotNone(house.owner_url)
+        self.assertIsNotNone(house.transfer_recipient)
+        self.assertIsNotNone(house.transferee_url)
+        self.assertIsNotNone(house.transfer_date)
+        self.assertIsNotNone(house.transfer_price)
+        self.assertTrue(house.transfer_accepted)
+
+    def test_house_parser_from_content_auction_with_bids(self):
+        content = self.load_resource(FILE_HOUSE_AUCTION_WITH_BIDS)
+
+        house = HouseParser.from_content(content)
+
+        self.assertIsNotNone(house)
         self.assertEqual(HouseStatus.AUCTIONED, house.status)
-        self.assertEqual(80_000, house.rent)
         self.assertIsNone(house.owner)
-        self.assertEqual(0, house.highest_bid)
+        self.assertIsNone(house.owner_url)
         self.assertIsNotNone(house.highest_bidder)
+        self.assertIsNotNone(house.highest_bid)
         self.assertIsNotNone(house.auction_end)
 
-    def test_house_parse_status_without_bids(self):
-        """Testing parsing the status of a house with no bids"""
-        house = HouseBuilder().name("Name")
-        content = self.load_resource(FILE_HOUSE_STATUS_NO_BIDS)
-        HouseParser._parse_status(house, content)
-        self.assertEqual(house._status, HouseStatus.AUCTIONED)
-        self.assertIsNone(house._auction_end)
+    def test_house_parser_from_content_auction_without_bids(self):
+        content = self.load_resource(FILE_HOUSE_AUCTION_WITHOUT_BIDS)
+
+        house = HouseParser.from_content(content)
+
+        self.assertIsNotNone(house)
+        self.assertEqual(HouseStatus.AUCTIONED, house.status)
+        self.assertIsNone(house.owner)
+        self.assertIsNone(house.owner_url)
+        self.assertIsNone(house.highest_bidder)
+        self.assertIsNone(house.highest_bid)
+        self.assertIsNone(house.auction_end)
 
     def test_house_parser_from_content_not_found(self):
         """Testing parsing a house that doesn't exist"""
@@ -94,6 +89,10 @@ class TestsHouse(TestCommons):
         content = self.load_resource(self.FILE_UNRELATED_SECTION)
         with self.assertRaises(InvalidContent):
             HouseParser.from_content(content)
+
+    # endregion
+
+    # region Houses Section Tests
 
     def test_houses_section_parser_from_content(self):
         """Testing parsing the house list of a world and city"""
@@ -118,6 +117,27 @@ class TestsHouse(TestCommons):
 
         self.assertEqual(houses[25].status, HouseStatus.RENTED)
 
+    def test_houses_section_parser_from_content_with_auctioned_houses(self):
+        """Testing parsing an empty house list"""
+        content = self.load_resource(FILE_HOUSE_SECTION_AUCTIONED_HOUSES)
+
+        houses_section = HousesSectionParser.from_content(content)
+
+        self.assertIsInstance(houses_section, HousesSection)
+
+        def test_house_entry_auction_started(house: HouseEntry):
+            self.assertEqual(HouseStatus.AUCTIONED, house.status)
+            self.assertIsNotNone(house.highest_bid)
+            self.assertIsNotNone(house.time_left)
+
+        def test_house_entry_auction_not_started(house: HouseEntry):
+            self.assertEqual(HouseStatus.AUCTIONED, house.status)
+            self.assertIsNone(house.highest_bid)
+            self.assertIsNone(house.time_left)
+
+        self.assertForAtLeastOne(houses_section.entries, test_house_entry_auction_not_started)
+        self.assertForAtLeastOne(houses_section.entries, test_house_entry_auction_started)
+
     def test_houses_section_parser_from_content_empty(self):
         """Testing parsing an empty house list"""
         content = self.load_resource(FILE_HOUSE_LIST_EMPTY)
@@ -130,7 +150,6 @@ class TestsHouse(TestCommons):
         self.assertEqual(HouseType.GUILDHALL, house_results.house_type)
         self.assertEqual(HouseOrder.RENT, house_results.order)
         self.assertIsEmpty(house_results.entries)
-
 
     def test_houses_section_parser_from_content_not_found(self):
         """Testing parsing an empty house list"""
@@ -145,3 +164,4 @@ class TestsHouse(TestCommons):
         with self.assertRaises(InvalidContent):
             HousesSectionParser.from_content(content)
 
+    # endregion