+++ tests/tests_tibiapy.py
@@ -1,11 +1,14 @@
+import contextlib
 import os.path
 import unittest
-from typing import Sized
+from typing import Callable, Iterable, Sized, TypeVar
 
 import tibiapy
 
 MY_PATH = os.path.abspath(os.path.dirname(__file__))
 RESOURCES_PATH = os.path.join(MY_PATH, "resources/")
+
+T = TypeVar("T")
 
 
 class TestCommons(unittest.TestCase):
@@ -22,6 +25,25 @@ class TestCommons(unittest.TestCase):
 
     def assertSizeAtLeast(self, collection: Sized, size: int, msg=None):
         self.assertGreaterEqual(len(collection), size, msg)
+
+    def assertForAtLeast(self, collection: Iterable[T], n: int, test: Callable[[T], None], msg=None):
+        pass_count = 0
+        for item in collection:
+            with contextlib.suppress(AssertionError):
+                test(item)
+                pass_count += 1
+        if pass_count < n:
+            self.fail(self._formatMessage(msg, f"Expected at least {n} to pass, {pass_count} passed."))
+
+    def assertForAll(self, collection: Iterable[T], test: Callable[[T], None], msg=None):
+        for item in collection:
+            try:
+                test(item)
+            except AssertionError as ae:
+                raise self.fail(self._formatMessage(msg, f"Expected all to pass: {ae}"))
+
+    def assertForAtLeastOne(self, collection: Iterable[T], test: Callable[[T], None], msg=None):
+        self.assertForAtLeast(collection, 1, test, msg)
 
     @staticmethod
     def load_resource(resource):